Switches to a Laplacian filter instead of a Sobel filter.  Resolves the rotation dependancy, performance still to be definitely determined.

+++ sensors/stereo/barrymoore.hpp
@@ -28,7 +28,7 @@ struct BarryMooreState
     int sadThreshold;
     
     // a parameter added to the SAD/sobel division in the denominator
-    int sobelAdd;
+    float sobelAdd;
     
     Mat mapxL;
